Quick unlock : Fix bug
There is an issue where passwords > 8 chars with quick unlock enabled
will be unable to unlock their lockscreen. This is due to the logic evaluating
anything beyond 3chars as a password attempt.

This change adds in a 2s delay after the last keypress before attempting a validation.
Once this change goes through, a cm7 backport will be made as well.

Change-Id: Ida6289c77f6724cb42e8d3b82593f29afe3d6d1b